fix(quality): decompose SettingsService — fragment merger, slug map, two reconcilers
SettingsService was 1356 lines (threshold 1000). Alongside the settings CRUD it
owned the ADR-037 fragment merge, the schema-slug data table, and both schema
reconcilers.

- RegisterFragmentMerger owns the ADR-037 deep-merge and the fragment-set hash
  that forces a re-import. It is a pure data transformation, so it is now
  instance-based and PUBLIC — the six fragment test suites that used to reach it
  through ReflectionMethod on a private static now just call it.
- SchemaSlugMap holds the slug -> appconfig-key table and the owned
  x-openregister-* annotation block names, plus the workflow_definition_schema
  alias constants that were previously repeated as string literals at two call
  sites.
- SchemaKeyReconciler owns both paths that write a *_schema key: the import
  result and the direct SchemaMapper slug lookup.
- SchemaAnnotationReconciler owns the declarative x-openregister-* merge onto
  live schemas.

Public API and the (appConfig, appManager, container, logger) constructor
signature are unchanged, so the bespoke factory and the ~180 injection sites are
untouched; the collaborators are constructed internally.

The pre-existing class-level @SuppressWarnings(PHPMD.ExcessiveClassComplexity)
is removed — verified by measuring with it gone.

phpmd: SettingsService ExcessiveClassLength 1356 -> cleared (1356 -> 810 lines);
repo total 18 -> 17 findings, no new finding anywhere in lib/.
