+++ tests/Unit/Settings/ZaakportaalFragmentTest.php
@@ -27,7 +27,7 @@ declare(strict_types=1);
 
 namespace OCA\Procest\Tests\Unit\Settings;
 
-use OCA\Procest\Service\SettingsService;
+use OCA\Procest\Service\Settings\RegisterFragmentMerger;
 use PHPUnit\Framework\TestCase;
 use ReflectionMethod;
 
@@ -56,12 +56,9 @@ class ZaakportaalFragmentTest extends TestCase
             true
         );
 
-        $reflection = new ReflectionMethod(SettingsService::class, 'mergeRegisterFragments');
-        $reflection->setAccessible(true);
-
-        [$merged] = $reflection->invokeArgs(
-            null,
-            [$base, __DIR__.'/../../../lib/Settings/register.d']
+        [$merged] = (new RegisterFragmentMerger())->merge(
+            base: $base,
+            fragmentDir: __DIR__.'/../../../lib/Settings/register.d'
         );
 
         $this->merged = $merged;
